Switch from Py_eval_input to Py_single_input
This flag is used by the interpreter to determine the context the
provided code is running in. `Py_single_input` is documented as "This
is the symbol used for the interactive interpreter loop.", meaning
this flag sets the compiled object's `kind` as `Interactive_kind`.

This is important if we want to use `sys.displayhook`, as the
displayhook is only executed for interactive code. This allows us to
do kinda a cool thing: by minimizing and adapting @rhelmot's
[monkeyhex][1] a little, we can make the repl print clickable hex by
default:

```
Python>list(Functions())
[1D001000,
 1D001020,
 1D001040,
 // ... snip
```

Additionally, this exposes the `_` variable to match a python repl: it
will store the result of the last executed command.

```
Python>65
65
Python>a = _
Python>a
65
```

Which is a minor quality of life improvement :)

[1]: https://github.com/rhelmot/monkeyhex/blob/master/monkeyhex.py