BUGFIX: IDAPython: upon failure, `ida_hexrays.decompile()` now always returns `None` (instead of either throw a `ida_hexrays.DecompilationFailure` or return `None`.)
The type `ida_hexrays.DecompilationFailure` is still present (but is a dummy) because otherwise scripts might break. It is never thrown anymore.

Note that `vds_xrefs.py` still need the `str(cfunc))` kludge, because of a caveat in when/how the decompiler actually updates the global xrefs cache. This will be fixed in the next IDA minor release

In addition, this commit introduces the notion of a typemap to convert form an `ea_t` to a `func_t *` for the `decompile()` call, but doesn't generalize its use.
Since this is a relatively far-ranging change, this will only be introduced in the next IDA minor release as well.